ui: improved, more configurable
SECTION_COLOR now colors all the section text, not just the name
Overhauled list formatting (with kv_pairs() )
added LIST_ACCENT_COLOR
hidden some redundant logs in the renaming audio files section
add track number and title tags to metadata preview list
changed method used for deleting cover arts (in debug.py), check if file exist instead of catching FileNotFoundError
add title case formatting options for kv_pairs() and new format presets
added comments in kv_pairs() definition

+++ debug.py
@@ -13,12 +13,9 @@ userdata = {
 def init(Data_obj):
     # remove square cover art
     coverartsqr = Data_obj.cover.get('square')
-    try:
+    if os.path.isfile(coverartsqr):
         os.remove(coverartsqr)
-        log.debug(f'Removed {coverartsqr}')
-    except FileNotFoundError: 
-        # if the file doesnt exist, no need to delete
-        pass
+        log.debug(f'Removed {filename(coverartsqr)}')
     del coverartsqr
 
     # remove a video if none is missing
